==> Main/rscripts/02_methodology.Rmd <==
Bank of International Settlement (BIS) defines two types of EER indexes (EER): Broad and narrow. The broad EER consists of 61 advanced and developing economies, whereas the narrow EER considers only 25 advanced economies. Although the broad basket is more representative than the narrow, neither should be regarded as the better measurement. The narrow indices may better gauge the competitiveness among advanced countries because their products have similar elasticities of substitution. The broad indices, on the other hand, give a more global picture by taking the emerging market economies into account. As a result, they would be more useful in analyses of issues such as the sustainability of the external trade balances. Our study is based on the broad EERs.

We retrieve broad nominal (available daily and monthly) and broad real (available monthly) effective exchange rate (EER) index data available from BIS for `r cname[1:(length(cname)-1)]` and `r cname[length(cname)]` from `r df_date[1]` to `r df_date[length(df_date)]`. These indices are base to 100 for year 2010. We deflate daily nominal EER by respective monthly deflator -which is ratio of monthly nominal and real effective exchange rate- to generate daily real EER (REER).

For the analysis, we developed two sets of data. First, we consider the are the growth rates of the weekly REER of each economy (see Figure 1) and second is the weekly volatility index (see Figure 2) based on the daily REER under the assumption of the volatility is fixed within the weekly period but variable across weeks following @Garman1980 and @Alizadeh2002 methodology. We consider the REER of Monday and Friday as weekly opening and closing value, and weekly minimum and maximum as the low and close value. The weekly estimated volatility is calculate as following:

\[{{\hat{\sigma }}^{2}}=0.511{{\left( {{H}_{t}}-{{L}_{t}} \right)}^{2}}-0.019\left[ \left( {{C}_{t}}-{{O}_{t}} \right)\left( {{H}_{t}}+{{L}_{t}}-2{{O}_{t}} \right)-2\left( {{H}_{t}}-{{O}_{t}} \right)\left( {{L}_{t}}-{{O}_{t}} \right) \right]-0.383{{\left( {{C}_{t}}-{{O}_{t}} \right)}^{2}}\]

where, $H$  is the Monday-Friday high, $L$  is the Monday-Friday low, $O$  is the Monday open and $C$  is the Friday close (all in natural logarithms).

### **Figure 1: Trend of weekly real effective exchange rate from `r df_date[1]` to `r df_date[length(df_date)]`.**

```{r, echo = FALSE, fig.width=9}
# rw <- read.csv('rscripts/outcomes/rw.csv')[,-1]
# rw_dat <- read.csv('rscripts/outcomes/rw_dat.csv')[,-1]
par(mfrow = c(2, 3), mai = c(0.25, 0.27, 0.25, 0.25))
for (i in 1:ncol(rw)){
  with(rw_dat, plot(Week, rw[,i], xaxt='n', type='l', main=as.matrix(colnames(rw))[i,1], xlab='', ylim=c(40,150)))
  axis.Date(1, at=seq(min(rw_dat$Week), max(rw_dat$Week), by='24 mon'), format='%Y')
  abline(h=100, lty = 2, col = 'grey')
}
```

### **Figure 2: Volatility of weekly real effective exchange rate from `r df_date[1]` to `r df_date[length(df_date)]`.**

```{r, echo=FALSE}
# wv <- read.csv('rscripts/outcomes/wv.csv')[ ,-1]
# wv_dat <- read.csv('rscripts/outcomes/wv_dat.csv')[,-1]
par(mfrow = c(2, 3), mai = c(0.25, 0.27, 0.25, 0.25))
for (i in 1:ncol(wv)){
  with(wv_dat, plot(Week, wv[,i], xaxt='n', type='l', main=as.matrix(colnames(wv))[i,1], xlab='', ylim=c(0,8)))
  axis.Date(1, at=seq(min(wv_dat$Week), max(wv_dat$Week), by='24 mon'), format='%Y')
  abline(h=100, lty = 2, col = 'grey')
}
```

## **Diebold and Yilmaz methodology**

### **Vector autoregression (VAR) model**

Vector autoregression (VAR) model is a stochastic process to capture the linear interdependencies among multiple stationary time series by incorporating own lagged values, the lagged values of the other model variables, and an error term. VAR modeling does not require as much knowledge about the forces influencing a variable as do structural models with simultaneous equations: The only prior knowledge required is a list of variables which can be hypothesized to affect each other intertemporally. Following Sims (1980), given $N$ stationary variables, a $p$ lagged vector autoregression $VAR\left( p \right)$ system can be defined as:

\[{{x}_{t}}=\sum\nolimits_{i=1}^{p}{{{\Phi }_{i}}{{x}_{t-i}}+{{\varepsilon }_{t}}}\]

The basic component of Diebold and Yilmaz (D&Y) model is the generalized variance decomposition of forecast errors of an $N-$ variables (economies), $p$ lagged covariance stationary vector auto-regression $VAR\left( p \right)$ approximating model. Such $VAR\left( p \right)$ system is defined as: 

\[{{x}_{t}}=\sum\nolimits_{i=1}^{p}{{{\Phi }_{i}}{{x}_{t-i}}+{{\varepsilon }_{t}}}\]

where ${{\varepsilon }_{t}}$ is the vectors independently and identically distributed disturbances and $\Omega$ is the covariance matrix. 
For a covariance stationary $VAR\left( p \right)$ process there exist the moving averages representation as ${{x}_{t}}=\sum\nolimits_{i=0}^{\infty }{{{A}_{i}}{{\varepsilon }_{t-i}}}$, where the $N\times N$ coefficient matrix ${{A}_{i}}$ obey the recursion ${{A}_{i}}={{\Phi }_{1}}{{A}_{i-1}}+{{\Phi }_{2}}{{A}_{i-2}}+\cdots +{{\Phi }_{p}}{{A}_{i-p}}$ with ${{A}_{0}}$ an $N\times N$ identity matrix and ${{A}_{i}}=0$ for $i<0$.The moving averages coefficients (or transformations such as impulse response functions or variance decomposition) provides the dynamics of the system. 

where, ${{\varepsilon }_{t}}$ is the vectors independently and identically distributed disturbances and $\Omega$ is the covariance matrix. For a covariance stationary $VAR\left( p \right)$ process there exist the moving averages representation as ${{x}_{t}}=\sum\nolimits_{i=0}^{\infty }{{{A}_{i}}{{\varepsilon }_{t-i}}}$, where the $N\times N$ coefficient matrix ${{A}_{i}}$ obey the recursion ${{A}_{i}}={{\Phi }_{1}}{{A}_{i-1}}+{{\Phi }_{2}}{{A}_{i-2}}+\cdots +{{\Phi }_{p}}{{A}_{i-p}}$ with ${{A}_{0}}$ being an $N\times N$ identity matrix and ${{A}_{i}}=0$ for $i<0$. The moving averages coefficients (or transformations such as impulse response functions  or variance decomposition) provides the dynamics of the system.

### **Variance decomposition matrix (VDM) and various connectedness**

The variance decompositions allow to split the $H$-step ahead forecast error of each variable into parts that can be attributable to the various market shocks. The aggregation of these decompositions will be subsequently used to compute the directional connectedness from an economy to any or to all the included economy.

The variance decompositions computation is usually done by using orthogonal VAR shocks. The Cholesky identification scheme achieves orthogonality but the computed variance decompositions are then unstable  and they are dependent on the ordering of the marketsmodel. Thus, Cholesky decomposition is not suitable. A framework that produces order-invariant decompositions is the generalized variance decompositions, of which it VAR that allows for correlated shocks but accounts for them appropriately. The framework, which we denote KPPS, has been first proposed by Koop et al. (1996) and Pesaran and Shin (1998). The KPPS forecast error variance decomposition matrix (VDM) for $H$-step ahead is computed as:

\[\theta _{ij}^{g}\left( H \right)=\frac{\sigma _{jj}^{-1}\sum\nolimits_{h=0}^{H-1}{{{\left( {{{{e}'}}_{i}}{{h}_{h}}\Omega {{e}_{j}} \right)}^{2}}}}{\sum\nolimits_{h=0}^{H-1}{\left( {{{{e}'}}_{i}}{{h}_{h}}\Omega {{e}_{j}} \right)}}\]

Where, ${{\sigma }_{jj}}$ is the standard deviation of the error term of the ${{j}^{th}}$ economy and ${{e}_{i}}$ is a selection vector with one on the ${{i}^{th}}$ element and zero otherwise.

The connectedness horizon $H$ is important because it helps to analyze the connectedness in "short-run" and "long run". Consider a pairwise connectedness example, shocks to $j$ many impact the forecast error variance of $i$ only with a lag, such that ${{C}_{i\leftarrow j}}$ may be small for small $H$ but nevertheless larger for larger $H$. Intuitively, as the horizon lengthens there may be more chances for connectedness to appear and $H$ approaches to infinity, it gives an unconditional variance decomposition. We use $H=10$ weeks in our main analysis but also provide results with different $H$ sequences. 

### **Pairwise direction connectedness**
To provide the information about the relative importance of each random innovation to the variables in the VAR (information relative to the size of endogenous and exogenous shocks on a given economy), we can normalize each entry with its respective row as:

\[\tilde{\theta }_{ij}^{g}\left( H \right)=\frac{\theta _{ij}^{g}\left( H \right)}{\sum\nolimits_{j=1}^{N}{\theta _{ij}^{g}\left( H \right)}}\]

Which implies the row sum for any ${{i}^{th}}$ economy to be $\sum\nolimits_{j=1}^{N}{\tilde{\theta }_{ij}^{g}\left( H \right)}=1$, hence row sums across all the economy be $\sum\nolimits_{i=1}^{N}{\sum\nolimits_{j=1}^{N}{\tilde{\theta }_{ij}^{g}\left( H \right)}}=N$.

Now each element of the row wise standardized VDM i.e ${\tilde \theta _{ij}^g\left( H \right)}$ can be interpreted as the pairwise directional connectedness from economy $j$ to economy $i$ at horizon $H$. For the intuitive purpose, ${{C}_{i\leftarrow j}}\left( H \right)$ represents the pairwise connectedness from economy $j$ to economy $i$ at horizon $H$ and the opposite direction pairwise connectedness from economy as ${{C}_{j\leftarrow i}}\left( H \right)$. The net pairwise directional connectedness then can be defined as

\[{{C}_{ij}}={{C}_{i\leftarrow j}}\left( H \right)-{{C}_{j\leftarrow i}}\left( H \right)\]

### **Net-pairwise directional connectedness**
The net pairwise directional connectedness identifies the dominant economy who transmit the information. Note that ${{C}_{ij}}=-{{C}_{ji}}$ and if ${{C}_{ij}}>0$ then, economy $j$transmit the information to economy $i$.

To find how all the economy are jointly contributing to a single economy, then the partial aggregation of connectedness from all economy to economy $i$ except from itself can be denoted as the row sum	

\[{{C}_{i\leftarrow \bullet }}\left( H \right)=\frac{\sum\nolimits_{j=1,j\ne i}^{N}{\tilde{\theta }_{ij}^{g}\left( H \right)}}{\sum\nolimits_{i=1}^{N}{\sum\nolimits_{j=1}^{N}{\tilde{\theta }_{ij}^{g}\left( H \right)}}}=\frac{\sum\nolimits_{j=1,j\ne i}^{N}{\tilde{\theta }_{ij}^{g}\left( H \right)}}{N}\] 

Similarly, to compute how a economy $i$ is contributing to the shocks of all other economy (except by itself) by aggregating partially as

\[{{C}_{\bullet \leftarrow i}}\left( H \right)=\frac{\sum\nolimits_{j=1,j\ne i}^{N}{\tilde{\theta }_{ji}^{g}\left( H \right)}}{\sum\nolimits_{i=1}^{N}{\sum\nolimits_{j=1}^{N}{\tilde{\theta }_{ij}^{g}\left( H \right)}}}=\frac{\sum\nolimits_{i=1,j\ne i}^{N}{\tilde{\theta }_{ji}^{g}\left( H \right)}}{N}\]

### **Net-directional connectedness and total connectedness**
Then the net directional connectedness can be measured as:

\[{{C}_{i}}\left( H \right)={{C}_{i\leftarrow \bullet }}\left( H \right)-{{C}_{\bullet \leftarrow i}}\left( H \right)\] 

The total aggregation of the variance decompositions across all economys measures the system-wide connectedness. The total connectedness in all economys can be computed as:

\[C\left( H \right)=\frac{\sum\nolimits_{i=1,i\ne j}^{N}{\sum\nolimits_{j=1,j\ne i}^{N}{\tilde{\theta }_{ji}^{g}\left( H \right)}}}{\sum\nolimits_{i=1}^{N}{\sum\nolimits_{j=1}^{N}{\tilde{\theta }_{ij}^{g}\left( H \right)}}}=\frac{\sum\nolimits_{i=1,i\ne j}^{N}{\sum\nolimits_{j=1,j\ne i}^{N}{\tilde{\theta }_{ji}^{g}\left( H \right)}}}{N}\]

### **Sample connectedness**
The static, unconditional connectedness can be expressed as $C\left( x,H,A\left( L \right) \right)$ because the connectedness $C$ depends upon the set of variables $x$ whose connectedness is to be examined, the predictive horizon $H$ for variance decompositions, and the dynamics $A\left( L \right)$. Since, $A\left( L \right)$ is unknown in reality but can be approximated using a finite ordered vector autoregression $M\left( L;\theta  \right)$ -which is a dynamic approximating model with finite dimensional parameter $\theta$. Allow for time-varying connectedness can equip analysis to capture evolution that may arise in economy from changing tastes, technologies, institutions, business cycles, abrupt financial market environments (e.g., crisis, non-crisis) [@Diebold2014]. Such time-varying connectedness can be expressed as ${{C}_{t}}\left( x,H,{{A}_{t}}\left( L \right),M\left( L;{{\theta }_{t}} \right) \right)$. Because of finite samples of observed data sample, we can only estimate approximating models, so we express the sample estimated connectedness as ${{\overset{\frown }{\mathop{C}}\,}_{t}}\left( x,H,{{A}_{t}}\left( L \right),M\left( L_{t}^{*};{{{\hat{\theta }}}_{t}} \right) \right)$ where the data sample runs from $t=1,\ldots ,T$ and $L_{t}^{*}$ is the optimum lag selection using some information criterion for approximating VAR model for each data sample. We use Akaike Information criterion to select the optimum lag for approximating VAR and allow 4 lags as the maximum lags. For this purpose, we use 250 weeks rolling window to perform the analysis and later facilitate several different rolling window periods for analysis.
